requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
volcengine-python-sdk[ark]
openai